MODTLR-191 Create two different event types
