+++ src/main/java/org/folio/support/KafkaEvent.java
@@ -18,53 +18,11 @@ import lombok.extern.log4j.Log4j2;
 @AllArgsConstructor
 @ToString
 public class KafkaEvent<T> {
-  private String id;
-  private String tenant;
-  private EventType type;
-  private long timestamp;
-  @ToString.Exclude
-  private EventData<T> data;
-
-  public KafkaEvent(String id, String tenant, EventType type, long timestamp, EventData<T> data,
-    String tenantIdHeaderValue) {
-
-    this.id = id;
-    this.tenant = tenant;
-    this.type = type;
-    this.timestamp = timestamp;
-    this.data = data;
-    this.tenantIdHeaderValue = tenantIdHeaderValue;
-  }
-
-  // For inventory topics
-  private String eventId;
-  @JsonProperty("old")
-  private T oldVersion;
-  @JsonProperty("new")
-  private T newVersion;
+  @With
+  @JsonIgnore
+  String tenantIdHeaderValue;
 
   @With
   @JsonIgnore
-  private String tenantIdHeaderValue;
-
-  @With
-  @JsonIgnore
-  private String userIdHeaderValue;
-
-  public enum EventType {
-    UPDATED, CREATED, DELETED, ALL_DELETED,
-    // For inventory topics
-    UPDATE, DELETE, CREATE, DELETE_ALL, REINDEX, ITERATE, MIGRATION
-  }
-
-  @Builder
-  @Getter
-  @NoArgsConstructor
-  @AllArgsConstructor
-  public static class EventData<T> {
-    @JsonProperty("old")
-    private T oldVersion;
-    @JsonProperty("new")
-    private T newVersion;
-  }
+  String userIdHeaderValue;
 }
